add utils file, fix lint errors
