+++ LICENSE
@@ -1,7 +1,6 @@
-Software License Agreement (BSD License)
-
-Copyright (c) 2014, Robert Bosch LLC.
 All rights reserved.
+
+Software License Agreement (BSD License 2.0)
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions
@@ -13,7 +12,7 @@ are met:
    copyright notice, this list of conditions and the following
    disclaimer in the documentation and/or other materials provided
    with the distribution.
- * Neither the name of the Robert Bosch LLC. nor the names of its
+ * Neither the name of the copyright holder nor the names of its
    contributors may be used to endorse or promote products derived
    from this software without specific prior written permission.
